backend/requirements.txt
polars>=0.20.0
pyarrow>=15.0.0
duckdb>=0.10.0
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.0
google-genai>=0.2.0
python-dotenv>=1.0.1
